Splits Layout into components
